requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Embedding models
transformers>=4.57.2
sentence-transformers>=3.0.0
torch>=2.0.0
datasets>=2.0.0
accelerate>=0.30.0

# Vector DB
qdrant-client>=1.12.0

# Document processing
pdf2image>=1.17.0
Pillow>=10.0.0

# OCR
paddlepaddle>=3.2.2
paddleocr>=3.4.0

# HWP document conversion
pyhwp>=0.1b15

# VLM API client
openai>=1.0.0

# Utilities
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
